Re-organise the feed classes in the core app
Separate the re-usable base classes from those that are actually
used to generate feeds.
