+++ hines/core/feeds/__init__.py
@@ -1,0 +1,10 @@
+from .base import ExtendedFeed, ExtendedRSSFeed
+from .feeds import EverythingFeedRSS, AdminWebmentionsFeedRSS
+
+
+__all___ = [
+    ExtendedFeed,
+    ExtendedRSSFeed,
+    EverythingFeedRSS,
+    AdminWebmentionsFeedRSS,
+]
